hephaestus/pyproject.toml
[project]
name = "hephaestus"
version = "0.1.0"
description = "HEPHAESTUS – PDF board-game component extraction system"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "HEPHAESTUS Team", email = "team@hephaestus.dev" }]
dependencies = [
    "pymupdf>=1.24.0",
    "Pillow>=10.0.0",
    "typer[all]>=0.12.0",
    "imagehash>=4.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "hypothesis>=6.0.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
hephaestus = "hephaestus.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"
log_level = "WARNING"
log_cli = false

[tool.mypy]
python_version = "3.11"
strict = false
plugins = []